backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlmodel==0.0.22
httpx==0.27.0
websockets==12.0
loguru==0.7.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
google-generativeai==0.7.2
